+++ Assets/GothicVR/Scripts/Creator/MultiResolutionMeshCreator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ab4010f6fe256f43aca77c2568f449f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
